Feature/game interaction

+++ src/main/java/com/novi/poffinhouse/dto/output/game/GameBerryOutputDto.java
@@ -6,10 +6,10 @@ import lombok.Data;
 @Data
 public class GameBerryOutputDto {
     private Long id;
-    private int indexNumber;
+    private Long indexNumber;
     private String name;
 
-    public GameBerryOutputDto(Long id, int indexNumber, @NotBlank String name) {
+    public GameBerryOutputDto(Long id, Long indexNumber, @NotBlank String name) {
         this.id = id;
         this.indexNumber = indexNumber;
         this.name = name;
